fix: 육류 등록 페이지 palette, 마진 적용
