+++ app/structure/lib/screen/meat_registration/creation_management_num_loading_screen.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter_spinkit/flutter_spinkit.dart';
+import 'package:structure/config/palette.dart';
 import 'package:structure/config/pallete.dart';
 
 class CreationManagementNumLoadingScreen extends StatelessWidget {
@@ -8,21 +9,31 @@ class CreationManagementNumLoadingScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Container(
-          margin: EdgeInsets.only(left: 46.w, top: 182.h),
-          child: Text('관리 번호를\n생성 중 입니다.', style: Pallete.h2),
-        ),
-        SizedBox(height: 80.h),
-        const SpinKitThreeBounce(
-          color: Pallete.mainBtnAtvBg,
-          size: 50.0,
-        ),
-        SizedBox(height: 90.h),
-        Center(child: Image.asset('assets/images/print.png')),
-      ],
+    return Container(
+      margin: EdgeInsets.symmetric(horizontal: 40.w),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Container(
+            margin: EdgeInsets.only(top: 188.h),
+            child: Text('관리 번호를\n생성 중 입니다', style: Pallete.h2),
+          ),
+          Expanded(
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.center,
+              mainAxisSize: MainAxisSize.min,
+              children: [
+                SpinKitThreeBounce(
+                  color: Palette.primary,
+                  size: 48.w,
+                ),
+                SizedBox(height: 88.h),
+                Center(child: Image.asset('assets/images/print.png')),
+              ],
+            ),
+          ),
+        ],
+      ),
     );
   }
 }
